flask-backend/requirements.txt
flask
tensorflow
numpy
pillow
joblib
scikit-learn